+++ src/Models/Image.php
@@ -2,14 +2,17 @@
 
 namespace DrewRoberts\Media\Models;
 
+
 use Illuminate\Database\Eloquent\Model;
+use Tipoff\Support\Models\BaseModel;
 use Tipoff\Support\Traits\HasCreator;
 use Tipoff\Support\Traits\HasPackageFactory;
 use Tipoff\Support\Traits\HasUpdater;
 
-class Image extends Model
+class Image extends BaseModel
 {
-    use HasPackageFactory, HasCreator, HasUpdater;
+
+    use HasCreator, HasUpdater, HasPackageFactory;
 
     protected $guarded = ['id'];