+++ src/Models/Tag.php
@@ -5,17 +5,17 @@ namespace DrewRoberts\Media\Models;
 use DrewRoberts\Media\Traits\HasSlug;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection as DbCollection;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use Spatie\EloquentSortable\Sortable;
 use Spatie\EloquentSortable\SortableTrait;
+use Tipoff\Support\Models\BaseModel;
 use Tipoff\Support\Traits\HasCreator;
 use Tipoff\Support\Traits\HasPackageFactory;
 use Tipoff\Support\Traits\HasUpdater;
 
-class Tag extends Model implements Sortable
+class Tag extends BaseModel implements Sortable
 {
-    use SortableTrait, HasSlug, HasPackageFactory, HasCreator, HasUpdater;
+    use SortableTrait, HasSlug, HasCreator, HasUpdater, HasPackageFactory;
 
     protected $guarded = ['id'];
 
@@ -106,5 +106,4 @@ class Tag extends Model implements Sortable
     {
         return static::groupBy('type')->pluck('type');
     }
-
 }